Refactor controller and scene switching
