+++ src/main/java/application/controller/SceneManager.java
@@ -1,0 +1,29 @@
+package application.controller;
+
+import application.utils.Paths;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
+import java.net.URL;
+
+public class SceneManager {
+    private Stage stage;
+
+    public SceneManager(Stage stage) {
+        this.stage = stage;
+    }
+
+    public void switchScene(String fxmlPath) {
+        try {
+            URL fxml = application.view.GUI.class.getResource(fxmlPath);
+            FXMLLoader loader = new FXMLLoader(fxml);
+            Scene scene = new Scene(loader.load());
+            stage.setScene(scene);
+            stage.show();
+        }
+        catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+}
