+++ src/main/java/application/utils/Paths.java
@@ -1,0 +1,7 @@
+package application.utils;
+
+public class Paths {
+    public static final String LOGIN = "/fxml/login.fxml";
+    public static final String APP = "/fxml/app.fxml";
+    public static final String POST = "/fxml/post_cell.fxml";
+}
