+++ src/main/java/application/view/GUI.java
@@ -1,19 +1,15 @@
 package application.view;
 
-import application.controller.Controller;
+import application.controller.SceneManager;
+import application.utils.Paths;
 import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
 import javafx.scene.image.Image;
 import javafx.stage.Stage;
 
-import java.net.URL;
 
 
 public class GUI extends Application {
-
-    private Stage stage;
-    private Controller sceneController;
+    private static SceneManager sceneManager;
 
     public void start(Stage stage) throws Exception {
         stage.setTitle("Shout!");
@@ -21,22 +17,11 @@ public class GUI extends Application {
         Image icon = new Image(getClass().getResourceAsStream("/images/icons/logo.png"));
         stage.getIcons().add(icon);
 
-        URL fxml = GUI.class.getResource("/fxml/login.fxml");
-        FXMLLoader loader = new FXMLLoader(fxml);
-        Scene scene = new Scene(loader.load());
-
-        stage.setScene(scene);
-        stage.show();
-
-        this.stage = stage;
+       sceneManager = new SceneManager(stage);
+       sceneManager.switchScene(Paths.LOGIN);
     }
 
-    @Override
-    public void init() {
-        sceneController = new Controller();
-    }
-
-    public Stage getStage() {
-        return stage;
+    public static SceneManager getSceneManager() {
+        return sceneManager;
     }
 }
